Add `babashka-project-tasks' command
This command is intended to be used with `project-switch-project', as
described in the docstring. In this context it has two advantages over
`babashka-tasks':

1. `babashka-tasks' will search for bb.edn files in
`default-directory' and `default-directory''s parent directories. This
means that when invoked by `project-switch-project' it will pick a
bb.edn file from the project of the current file you're editing,
instead of the project you're switching to.

2. When running `project-switch-project' and selecting the Babashka
task execution command, it is undesireable that the command recurses
into parent directories if there is no bb.edn file in the root
directory of the project.